+++ Assets/Razeware/Scripts/IDamageReceiver.cs
@@ -4,5 +4,5 @@ public interface IDamageReceiver
 {
     public event Action OnTakeDamage;
     public event Action OnDeath;
-    public void TakeDamage(int value);
+    public void TakeDamage(float value);
 }
